pyproject.toml
[project]
name = "porchsongs"
version = "1.0.0"
description = "Personal song lyric rewriter — make every song yours"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn[standard]==0.34.0",
    "sqlalchemy==2.0.36",
    "pydantic==2.10.4",
    "pydantic-settings==2.7.1",
    "python-dotenv==1.0.1",
    "requests==2.32.3",
    "beautifulsoup4==4.12.3",
    "any-llm-sdk>=1.8.0",
    "curl-cffi>=0.14.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "ruff>=0.11",
    "ty>=0.0.17",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend/app"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["backend"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "ANN", "RUF"]
ignore = [
    "E501",   # line too long — handled by formatter
    "B008",   # Depends() in function defaults — standard FastAPI pattern
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["backend"]

[tool.ty]

[tool.ty.environment]
python-version = "3.11"
python = ".venv"
root = ["backend"]
